+++ pkg/services/object/acl/v2/service.go
@@ -1,15 +1,19 @@
 package v2
 
 import (
+	"crypto/ecdsa"
 	"crypto/sha256"
 	"errors"
 	"fmt"
+	"time"
 
 	lru "github.com/hashicorp/golang-lru/v2"
 	"github.com/nspcc-dev/neo-go/pkg/core/block"
 	"github.com/nspcc-dev/neo-go/pkg/core/transaction"
 	"github.com/nspcc-dev/neo-go/pkg/neorpc/result"
 	"github.com/nspcc-dev/neo-go/pkg/smartcontract/trigger"
+	"github.com/nspcc-dev/neo-go/pkg/util"
+	"github.com/nspcc-dev/neofs-node/internal/chaintime"
 	icrypto "github.com/nspcc-dev/neofs-node/internal/crypto"
 	"github.com/nspcc-dev/neofs-node/pkg/core/container"
 	"github.com/nspcc-dev/neofs-node/pkg/core/netmap"
@@ -17,18 +21,21 @@ import (
 	apistatus "github.com/nspcc-dev/neofs-sdk-go/client/status"
 	"github.com/nspcc-dev/neofs-sdk-go/container/acl"
 	cid "github.com/nspcc-dev/neofs-sdk-go/container/id"
+	"github.com/nspcc-dev/neofs-sdk-go/object"
 	oid "github.com/nspcc-dev/neofs-sdk-go/object/id"
 	protoacl "github.com/nspcc-dev/neofs-sdk-go/proto/acl"
 	protoobject "github.com/nspcc-dev/neofs-sdk-go/proto/object"
 	protosession "github.com/nspcc-dev/neofs-sdk-go/proto/session"
 	sessionSDK "github.com/nspcc-dev/neofs-sdk-go/session"
+	sessionSDKv2 "github.com/nspcc-dev/neofs-sdk-go/session/v2"
 	"github.com/nspcc-dev/neofs-sdk-go/user"
 	"go.uber.org/zap"
 )
 
 type sessionTokenCommonCheckResult struct {
-	token sessionSDK.Object
-	err   error
+	token   sessionSDK.Object
+	tokenV2 sessionSDKv2.Token
+	err     error
 }
 
 type bearerTokenCommonCheckResult struct {
@@ -36,11 +43,55 @@ type bearerTokenCommonCheckResult struct {
 	err   error
 }
 
+type verifiedSession interface {
+	AuthorID() user.ID
+	AuthorKey() []byte
+}
+
+type verifiedObjectSessionV1 struct {
+	author user.ID
+	key    []byte
+}
+
+func (s verifiedObjectSessionV1) AuthorID() user.ID { return s.author }
+
+func (s verifiedObjectSessionV1) AuthorKey() []byte { return s.key }
+
+type verifiedObjectSessionV2 struct {
+	author user.ID
+	key    []byte
+}
+
+func (s verifiedObjectSessionV2) AuthorID() user.ID { return s.author }
+
+func (s verifiedObjectSessionV2) AuthorKey() []byte { return s.key }
+
+type nnsResolver struct {
+	fsChain FSChain
+
+	nnsCache *lru.Cache[string, bool]
+}
+
+func (r nnsResolver) HasUser(name string, userID user.ID) (bool, error) {
+	cacheKey := name + userID.String()
+	hasUser, ok := r.nnsCache.Get(cacheKey)
+	if ok {
+		return hasUser, nil
+	}
+	hasUser, err := r.fsChain.HasUserInNNS(name, userID.ScriptHash())
+	if err != nil {
+		return false, err
+	}
+	r.nnsCache.Add(cacheKey, hasUser)
+	return hasUser, nil
+}
+
 // Service checks basic ACL rules.
 type Service struct {
 	*cfg
 
 	c senderClassifier
+	r nnsResolver
 
 	sessionTokenCommonCheckCache *lru.Cache[[sha256.Size]byte, sessionTokenCommonCheckResult]
 	bearerTokenCommonCheckCache  *lru.Cache[[sha256.Size]byte, bearerTokenCommonCheckResult]
@@ -58,6 +109,18 @@ type FSChain interface {
 	// from the referenced container either in the current or the previous NeoFS
 	// epoch.
 	InContainerInLastTwoEpochs(_ cid.ID, pub []byte) (bool, error)
+
+	// HasUserInNNS checks whether given user is listed in the NNS domain.
+	HasUserInNNS(name string, addr util.Uint160) (bool, error)
+
+	// GetBlockHeader returns FS chain block header by its index.
+	GetBlockHeader(ind uint32) (*block.Header, error)
+
+	// BlockCount returns current FS chain block count.
+	BlockCount() (res uint32, err error)
+
+	// MsPerBlock returns milliseconds per block protocol configuration.
+	MsPerBlock() (res int64, err error)
 }
 
 // Netmapper must provide network map information.
@@ -79,6 +142,17 @@ type cfg struct {
 	irFetcher InnerRingFetcher
 
 	nm Netmapper
+
+	nodeKey *ecdsa.PublicKey
+
+	chainTime chaintime.TimeProvider
+}
+
+// WithTimeProvider sets external chain time provider updated from header subscriptions.
+func WithTimeProvider(p chaintime.TimeProvider) Option {
+	return func(c *cfg) {
+		c.chainTime = p
+	}
 }
 
 func defaultCfg() *cfg {
@@ -115,12 +189,21 @@ func New(fsChain FSChain, opts ...Option) Service {
 		panic(fmt.Errorf("unexpected error in lru.New: %w", err))
 	}
 
+	nnsCache, err := lru.New[string, bool](100)
+	if err != nil {
+		panic(fmt.Errorf("unexpected error in lru.New: %w", err))
+	}
+
 	return Service{
 		cfg: cfg,
 		c: senderClassifier{
 			log:       cfg.log,
 			innerRing: cfg.irFetcher,
 			fsChain:   fsChain,
+		},
+		r: nnsResolver{
+			fsChain:  fsChain,
+			nnsCache: nnsCache,
 		},
 		sessionTokenCommonCheckCache: sessionTokenCheckCache,
 		bearerTokenCommonCheckCache:  bearerTokenCheckCache,
@@ -131,12 +214,21 @@ func New(fsChain FSChain, opts ...Option) Service {
 func (b Service) ResetTokenCheckCache() {
 	b.sessionTokenCommonCheckCache.Purge()
 	b.bearerTokenCommonCheckCache.Purge()
+	b.r.nnsCache.Purge()
 }
 
-func (b Service) getVerifiedSessionToken(mh *protosession.RequestMetaHeader, reqVerb sessionSDK.ObjectVerb, reqCnr cid.ID, reqObj oid.ID) (*sessionSDK.Object, error) {
+func (b Service) getVerifiedSessionToken(mh *protosession.RequestMetaHeader, reqVerb sessionSDK.ObjectVerb,
+	reqVerbV2 sessionSDKv2.Verb, reqCnr cid.ID, reqObj oid.ID) (verifiedSession, error) {
 	for omh := mh.GetOrigin(); omh != nil; omh = mh.GetOrigin() {
 		mh = omh
 	}
+
+	mV2 := mh.GetSessionTokenV2()
+	if mV2 != nil {
+		return b.getVerifiedSessionTokenV2(mV2, reqVerbV2, reqCnr, reqObj)
+	}
+
+	// Fall back to V1 token
 	m := mh.GetSessionToken()
 	if m == nil {
 		return nil, nil
@@ -160,7 +252,15 @@ func (b Service) getVerifiedSessionToken(mh *protosession.RequestMetaHeader, req
 		return nil, err
 	}
 
-	return &res.token, nil
+	sig, ok := res.token.Signature()
+	if !ok {
+		return nil, errors.New("missing signature in session token")
+	}
+
+	return verifiedObjectSessionV1{
+		author: res.token.Issuer(),
+		key:    sig.PublicKeyBytes(),
+	}, nil
 }
 
 func (b Service) decodeAndVerifySessionTokenCommon(m *protosession.SessionToken) (sessionSDK.Object, error) {
@@ -200,6 +300,116 @@ func (b Service) verifySessionTokenAgainstRequest(token sessionSDK.Object, reqVe
 	}
 
 	return nil
+}
+
+// getVerifiedSessionTokenV2 validates and returns V2 session token info.
+func (b Service) getVerifiedSessionTokenV2(mV2 *protosession.SessionTokenV2, reqVerb sessionSDKv2.Verb, reqCnr cid.ID, reqObj oid.ID) (verifiedSession, error) {
+	mb := make([]byte, mV2.MarshaledSize())
+	mV2.MarshalStable(mb)
+
+	cacheKey := sha256.Sum256(mb)
+	res, ok := b.sessionTokenCommonCheckCache.Get(cacheKey)
+	if !ok {
+		res.tokenV2, res.err = b.decodeAndVerifySessionTokenV2Common(mV2)
+		b.sessionTokenCommonCheckCache.Add(cacheKey, res)
+	}
+	if res.err != nil {
+		return nil, res.err
+	}
+
+	if err := b.verifySessionTokenV2AgainstRequest(res.tokenV2, reqVerb, reqCnr, reqObj); err != nil {
+		return nil, err
+	}
+
+	var key []byte
+	origin := &res.tokenV2
+	for origin != nil {
+		sig, ok := origin.Signature()
+		if !ok {
+			return nil, errors.New("missing signature in V2 session token")
+		}
+		key = sig.PublicKeyBytes()
+		origin = origin.Origin()
+	}
+
+	return verifiedObjectSessionV2{
+		author: res.tokenV2.OriginalIssuer(),
+		key:    key,
+	}, nil
+}
+
+func (b Service) decodeAndVerifySessionTokenV2Common(m *protosession.SessionTokenV2) (sessionSDKv2.Token, error) {
+	var token sessionSDKv2.Token
+	if err := token.FromProtoMessage(m); err != nil {
+		return token, fmt.Errorf("invalid V2 session token: %w", err)
+	}
+
+	if err := token.ValidateWithNNS(b.r); err != nil {
+		return token, fmt.Errorf("invalid V2 session token: %w", err)
+	}
+
+	if b.nodeKey != nil {
+		ok, err := token.AssertAuthority(user.NewFromECDSAPublicKey(*b.nodeKey), b.r)
+		if err != nil {
+			return token, fmt.Errorf("v2 token authority check failed: %w", err)
+		}
+		if !ok {
+			return token, errors.New("v2 token doesn't authorize this node")
+		}
+	}
+
+	currentTime, err := b.currentChainTime()
+	if err != nil {
+		return token, err
+	}
+	currentTime = currentTime.Round(time.Second)
+	if !token.ValidAt(currentTime) {
+		return token, fmt.Errorf("%s: V2 token is invalid at %s, token iat %s, nbf %s, exp %s", invalidRequestMessage, currentTime, token.Iat(), token.Nbf(), token.Exp())
+	}
+
+	return token, nil
+}
+
+func (b Service) verifySessionTokenV2AgainstRequest(token sessionSDKv2.Token, reqVerb sessionSDKv2.Verb, reqCnr cid.ID, reqObj oid.ID) error {
+	if !reqObj.IsZero() {
+		if !token.AssertObject(reqVerb, reqCnr, reqObj) {
+			return errors.New("v2 session token does not authorize access to the object")
+		}
+	} else {
+		if !token.AssertVerb(reqVerb, reqCnr) {
+			return errInvalidVerb
+		}
+	}
+
+	return nil
+}
+
+// currentChainTime returns chain time preferring external provider.
+// If provider is not set, falls back to contract latest block header.
+func (b Service) currentChainTime() (time.Time, error) {
+	blockTimeMs, err := b.c.fsChain.MsPerBlock()
+	if err != nil {
+		return time.Time{}, fmt.Errorf("get ms per block protocol config: %w", err)
+	}
+	durBlockTime := time.Duration(blockTimeMs) * time.Millisecond
+	if b.chainTime != nil {
+		ct := b.chainTime.Now()
+		if !ct.IsZero() {
+			return ct.Add(durBlockTime), nil
+		}
+		// If provider returns zero, fall back to chain.
+	}
+
+	idx, err := b.c.fsChain.BlockCount()
+	if err != nil {
+		return time.Time{}, fmt.Errorf("get FS chain block count: %w", err)
+	}
+
+	header, err := b.c.fsChain.GetBlockHeader(idx - 1)
+	if err != nil {
+		return time.Time{}, fmt.Errorf("get FS chain block header: %w", err)
+	}
+	return time.UnixMilli(int64(header.Timestamp)).Add(durBlockTime), nil
 }
 
 func (b Service) getVerifiedBearerToken(mh *protosession.RequestMetaHeader, reqCnr cid.ID, ownerCnr user.ID, usrSender user.ID) (*bearer.Token, error) {
@@ -294,7 +504,7 @@ func (b Service) GetRequestToInfo(request *protoobject.GetRequest) (RequestInfo,
 		return RequestInfo{}, err
 	}
 
-	return b.findRequestInfo(request, cnr, acl.OpObjectGet, sessionSDK.VerbObjectGet, *obj)
+	return b.findRequestInfo(request, cnr, acl.OpObjectGet, sessionSDK.VerbObjectGet, sessionSDKv2.VerbObjectGet, *obj)
 }
 
 // HeadRequestToInfo resolves RequestInfo from the request to check it using
@@ -310,7 +520,7 @@ func (b Service) HeadRequestToInfo(request *protoobject.HeadRequest) (RequestInf
 		return RequestInfo{}, err
 	}
 
-	return b.findRequestInfo(request, cnr, acl.OpObjectHead, sessionSDK.VerbObjectHead, *obj)
+	return b.findRequestInfo(request, cnr, acl.OpObjectHead, sessionSDK.VerbObjectHead, sessionSDKv2.VerbObjectHead, *obj)
 }
 
 // SearchRequestToInfo resolves RequestInfo from the request to check it using
@@ -335,7 +545,7 @@ func (b Service) searchRequestToInfo(request interface {
 		return RequestInfo{}, err
 	}
 
-	return b.findRequestInfo(request, id, acl.OpObjectSearch, sessionSDK.VerbObjectSearch, oid.ID{})
+	return b.findRequestInfo(request, id, acl.OpObjectSearch, sessionSDK.VerbObjectSearch, sessionSDKv2.VerbObjectSearch, oid.ID{})
 }
 
 // DeleteRequestToInfo resolves RequestInfo from the request to check it using
@@ -351,7 +561,7 @@ func (b Service) DeleteRequestToInfo(request *protoobject.DeleteRequest) (Reques
 		return RequestInfo{}, err
 	}
 
-	return b.findRequestInfo(request, cnr, acl.OpObjectDelete, sessionSDK.VerbObjectDelete, *obj)
+	return b.findRequestInfo(request, cnr, acl.OpObjectDelete, sessionSDK.VerbObjectDelete, sessionSDKv2.VerbObjectDelete, *obj)
 }
 
 // RangeRequestToInfo resolves RequestInfo from the request to check it using
@@ -367,7 +577,7 @@ func (b Service) RangeRequestToInfo(request *protoobject.GetRangeRequest) (Reque
 		return RequestInfo{}, err
 	}
 
-	return b.findRequestInfo(request, cnr, acl.OpObjectRange, sessionSDK.VerbObjectRange, *obj)
+	return b.findRequestInfo(request, cnr, acl.OpObjectRange, sessionSDK.VerbObjectRange, sessionSDKv2.VerbObjectRange, *obj)
 }
 
 // HashRequestToInfo resolves RequestInfo from the request to check it using
@@ -383,7 +593,7 @@ func (b Service) HashRequestToInfo(request *protoobject.GetRangeHashRequest) (Re
 		return RequestInfo{}, err
 	}
 
-	return b.findRequestInfo(request, cnr, acl.OpObjectHash, sessionSDK.VerbObjectRangeHash, *obj)
+	return b.findRequestInfo(request, cnr, acl.OpObjectHash, sessionSDK.VerbObjectRangeHash, sessionSDKv2.VerbObjectRangeHash, *obj)
 }
 
 var ErrSkipRequest = errors.New("skip request")
@@ -450,15 +660,21 @@ func (b Service) PutRequestToInfo(request *protoobject.PutRequest) (RequestInfo,
 		}
 	}
 
-	op, verb := acl.OpObjectPut, sessionSDK.VerbObjectPut
+	op, verb, verbv2 := acl.OpObjectPut, sessionSDK.VerbObjectPut, sessionSDKv2.VerbObjectPut
 	tombstone := header.GetObjectType() == protoobject.ObjectType_TOMBSTONE
 	if tombstone {
 		// such objects are specific - saving them is essentially the removal of other
 		// objects
-		op, verb = acl.OpObjectDelete, sessionSDK.VerbObjectDelete
+		op, verb, verbv2 = acl.OpObjectDelete, sessionSDK.VerbObjectDelete, sessionSDKv2.VerbObjectDelete
+		for _, attr := range header.Attributes {
+			if attr.Key == object.AttributeAssociatedObject {
+				_ = obj.DecodeString(attr.Value)
+				break
+			}
+		}
 	}
 
-	reqInfo, err := b.findRequestInfo(request, cnr, op, verb, obj)
+	reqInfo, err := b.findRequestInfo(request, cnr, op, verb, verbv2, obj)
 	if err != nil {
 		return RequestInfo{}, user.ID{}, err
 	}
@@ -479,12 +695,13 @@ func (b Service) PutRequestToInfo(request *protoobject.PutRequest) (RequestInfo,
 func (b Service) findRequestInfo(req interface {
 	GetMetaHeader() *protosession.RequestMetaHeader
 	GetVerifyHeader() *protosession.RequestVerificationHeader
-}, idCnr cid.ID, op acl.Op, verb sessionSDK.ObjectVerb, obj oid.ID) (RequestInfo, error) {
+}, idCnr cid.ID, op acl.Op, verb sessionSDK.ObjectVerb, verb2 sessionSDKv2.Verb, obj oid.ID) (RequestInfo, error) {
 	var (
-		info    RequestInfo
-		metaHdr = req.GetMetaHeader()
+		info      RequestInfo
+		metaHdr   = req.GetMetaHeader()
+		verifyHdr = req.GetVerifyHeader()
 	)
-	sTok, err := b.getVerifiedSessionToken(metaHdr, verb, idCnr, obj)
+	sTok, err := b.getVerifiedSessionToken(metaHdr, verb, verb2, idCnr, obj)
 	if err != nil {
 		return info, err
 	}
@@ -492,14 +709,10 @@ func (b Service) findRequestInfo(req interface {
 	var reqAuthor user.ID
 	var reqAuthorPub []byte
 	if sTok != nil {
-		reqAuthor = sTok.Issuer()
-		sig, ok := sTok.Signature()
-		if !ok {
-			return info, errors.New("missing signature in session token")
-		}
-		reqAuthorPub = sig.PublicKeyBytes()
+		reqAuthor = sTok.AuthorID()
+		reqAuthorPub = sTok.AuthorKey()
 	} else {
-		if reqAuthor, reqAuthorPub, err = icrypto.GetRequestAuthor(req.GetVerifyHeader()); err != nil {
+		if reqAuthor, reqAuthorPub, err = icrypto.GetRequestAuthor(verifyHdr); err != nil {
 			return info, fmt.Errorf("get request author: %w", err)
 		}
 	}