+++ pkg/services/object/acl/v2/service_test.go
@@ -7,6 +7,7 @@ import (
 	"github.com/nspcc-dev/neo-go/pkg/core/transaction"
 	"github.com/nspcc-dev/neo-go/pkg/neorpc/result"
 	"github.com/nspcc-dev/neo-go/pkg/smartcontract/trigger"
+	"github.com/nspcc-dev/neo-go/pkg/util"
 	aclsvc "github.com/nspcc-dev/neofs-node/pkg/services/object/acl/v2"
 	"github.com/nspcc-dev/neofs-sdk-go/bearer"
 	apistatus "github.com/nspcc-dev/neofs-sdk-go/client/status"
@@ -26,12 +27,28 @@ import (
 
 type mockFSChain struct{}
 
+func (x *mockFSChain) MsPerBlock() (res int64, err error) {
+	panic("unimplemented")
+}
+
 func (x *mockFSChain) InvokeContainedScript(*transaction.Transaction, *block.Header, *trigger.Type, *bool) (*result.Invoke, error) {
 	panic("unimplemented")
 }
 
 func (x *mockFSChain) InContainerInLastTwoEpochs(cid.ID, []byte) (bool, error) {
 	return false, nil
+}
+
+func (x *mockFSChain) HasUserInNNS(string, util.Uint160) (bool, error) {
+	panic("unimplemented")
+}
+
+func (x *mockFSChain) GetBlockHeader(uint32) (*block.Header, error) {
+	panic("unimplemented")
+}
+
+func (x *mockFSChain) BlockCount() (uint32, error) {
+	panic("unimplemented")
 }
 
 type mockIR struct {