+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Changelog for NeoFS Node
 - `neofs-adm balance container-status` command (#3693)
 - IR supports new session token v2 (#3671)
 - SN supports new session token v2 for container operations (#3671)
+- SN supports new session token v2 for object operations (#3671)
 
 ### Fixed
 - IR panics at graceful shutdown (#3706)